Create lib.rs and rename lib folder to commons
