+++ src/endpoints/categories.rs
@@ -5,9 +5,9 @@ extern crate rocket_contrib;
 use rocket_contrib::Json;
 use diesel::prelude::*;
 use data::category::Category;
-use lib::pg_pool::DbConn;
+use commons::pg_pool::DbConn;
 use data::schema::categories::dsl::*;
-use lib::responders::*;
+use commons::responders::*;
 
 #[get("/categories")]
 pub fn get_categories(conn: DbConn) -> QueryResult<Cached<Cors<Json<Vec<Category>>>>> {
